refactor: 도메인별(locker) 패키지 분리 및 DTO 이름 수정
